RAG/requirements.txt
flask
pandas
psycopg2-binary
sentence-transformers
chromadb
requests
